requirements.txt
j2cli
jinja2
jq
openpyxl
pandas
pyyaml
yq
